Переименовывает 'CatalogGenresItem' в 'GenresItem'.
